feat - separate Search page components
