+++ components/Search/organisms/SearchBody/style.module.scss
